+++ .old/components/Entry/IndividualEntry.tsx
