+++ .old/components/ImageCapture.tsx
