+++ .old/components/Profile/Inventory.tsx
