+++ .old/components/Scrapbooking/ScrapbookCanvas.tsx
