+++ components/Themed.tsx
@@ -6,7 +6,7 @@
 import React from "react";
 import { Text as DefaultText, View as DefaultView } from "react-native";
 
-import { useColorScheme } from "./useColorScheme";
+import { useColorScheme } from "react-native";
 
 import Colors from "@/constants/Colors";
